pyproject.toml
[build-system]
requires = [
    "setuptools>=66",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "djangocms-ranked-search"
version = "0.1.0"
description = "Ranked search integration for django CMS using Haystack with accent-aware scoring."
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.8"
keywords = ["django", "django-cms", "djangocms", "search", "haystack", "whoosh"]
dependencies = [
    "Django>=3.2",
    "django-cms>=3.11",
    "django-haystack>=3.2",
    "aldryn-search>=2.0",
    "Whoosh>=2.7",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3",
    "Framework :: Django :: 4",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.optional-dependencies]
xapian = [
    "haystack-xapian>=1.2.3",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["djangocms_ranked_search"]
exclude = ["tests*", "docs*"]
